Differential Count Analysis with a Multinomial Topic Model

Implements methods for analysis of differential count expression using a topic model. These methods are motivated by gene expression studies, but could have other uses, such as identiying “key words” in topics derived from text documents.

diff_count_analysis(
  fit,
  X,
  s = rowSums(X),
  numiter = 100,
  tol = 1e-08,
  e = 1e-15,
  verbose = TRUE
)

Arguments

fit
An object of class “poisson_nmf_fit” or “multinom_topic_model_fit”. If a Poisson NMF fit is provided as input, the corresponding multinomial topic model fit is automatically recovered using link{poisson2multinom}.

X
The n x m counts matrix. It can be a sparse matrix (class "dgCMatrix") or dense matrix (class "matrix").

s
A numeric vector of length n determining how the rates are scaled in the Poisson model. See “Details” for guidance on the choice of s.

numiter
The number of EM updates performed to compute the maximum-likelihood estimates of the Poisson model parameters.

tol
When tol > 0, the EM algorithm for computing maximum-likelihood estimates will be stopped early when the largest change between two successive updates is less than tol.

e
A small, positive scalar included in some computations to avoid logarithms of zero and division by zero.

verbose
When verbose = TRUE, progress information is printed to the console.

Value

The return value is a list with six m x k matrices, where m is the number of columns in the counts matrix, and k is the number of topics, and an additional vector:

colmeans

A vector of length m containing the count averages (colMeans(X)).

F0

Estimates of the Poisson model parameters \(f_0\).

F1

Estimates of the Poisson model parameters \(f_1\).

beta

Log-fold change estimates, beta = log2(F1/F0).

se

Standard errors for the log-fold change estimates.

Z

Log-fold change z-scores.

pval

-log10 two-tailed p-values computed from the z-scores.

Details

The methods are based on the following univariate (“single-count”) Poisson model: $$x_i ~ Poisson(s_i \lambda_i),$$ in which the Poisson rates are defined as $$\lambda_i = (1 - q_i) f_0 + q_i f_1,$$ and where the two unknowns to be estimated are \(f_0, f_1 > 0.\) This model is applied separately to each count (column of the counts matrix), and to each topic k. The \(q_i\)'s are the topic probabilities for a given topic. An EM algorithm is used to compute maximum-likelihood estimates (MLEs) of the two unknowns.

The log-fold change statistics are defined as the log-ratio of the two Poisson rate parameters, \(\beta = \log_2(f_1/f_0)\). This statistic measures the increase (or decrease) in occurence in one topic compared to all other topics. The use of the base-2 logarithm comes from the convention used in gene expression studies.

When each \(s_i\) (specified by input argument s) is equal the total count for sample i (this is the default setting in diff_count_analysis), the Poisson model will closely approximate a binomial model of the count data, so that the Poisson model parameters \(f_0, f_1\) represent binomial probabilities. In this case, \(\beta\) represents the log-fold change in relative occurrence. (This Poisson approximation to the binomial is most accurate when the total counts rowSums(X) are large and \(f_0, f_1\) are small.)

Other choices for s are possible, and implement different normalization schemes for the counts, and different interpretations of the log-fold change statistics. Setting s to all ones implements the differential count analysis with no normalization, and \(\beta\) represents the log-fold change in absolute occurrence. This choice could be appropriate in settings where the total count is well-controlled across samples (rows of X).

The standard error and z-score calculations are based on a Laplace approximation to the likelihood at the MLE. This is the same strategy used to compute standard errors and z-scores in glm.