Droplet single-cell RNA-seq read count data from Montoro et al (2018)

These data are gene expression profiles of trachea epithelial cells in C57BL/6 mice obtained using droplet-based 3' single-cell RNA-seq. They were prepared from file GSE103354_Trachea_droplet_UMIcounts.txt.gz downloaded from the Gene Expression Omnibus (GEO) website, accession GSE103354.

Format

droplet is a 7,193 x 17,133 sparse matrix of read counts, with rows corresponding to samples (cells), and columns corresponding to genes.

References

D. T. Montoro et al (2018). A revised airway epithelial hierarchy includes CFTR-expressing ionocytes. Nature 560, 319–-324.

Examples

# Roughly 10% of the read counts are greater than zero.
data(droplet)
nnzero(droplet)/length(droplet)

#> [1] 0.099922