PCA, t-SNE and UMAP Plots

Visualize the structure of the Poisson NMF loadings or the multinomial topic model topic proportions by projection onto a 2-d surface. plot_hexbin_plot is most useful for visualizing the PCs of a data set with thousands of samples or more.

embedding_plot_2d(
  fit,
  Y,
  fill = "loading",
  k,
  fill.label,
  ggplot_call = embedding_plot_2d_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots)
)

embedding_plot_2d_ggplot_call(
  Y,
  fill,
  fill.type = c("loading", "numeric", "factor", "none"),
  fill.label,
  font.size = 9
)

pca_plot(
  fit,
  Y,
  pcs = 1:2,
  n = 10000,
  fill = "loading",
  k,
  fill.label,
  ggplot_call = embedding_plot_2d_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)

tsne_plot(
  fit,
  Y,
  n = 2000,
  fill = "loading",
  k,
  fill.label,
  ggplot_call = embedding_plot_2d_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)

umap_plot(
  fit,
  Y,
  n = 2000,
  fill = "loading",
  k,
  fill.label,
  ggplot_call = embedding_plot_2d_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)

pca_hexbin_plot(
  fit,
  Y,
  pcs = 1:2,
  bins = 40,
  breaks = c(0, 1, 10, 100, 1000, Inf),
  ggplot_call = pca_hexbin_plot_ggplot_call,
  ...
)

pca_hexbin_plot_ggplot_call(Y, bins, breaks, font.size = 9)

Arguments

fit

An object of class “poisson_nmf_fit” or “multinom_topic_model_fit”.

Y

The n x 2 matrix containing the 2-d embedding, where n is the number of rows in fit$L. If not provided, the embedding will be computed automatically.

fill

The quantity to map onto the fill colour of the points in the PCA plot. Set fill = "loading" to vary the fill colour according to the loadings (or topic proportions) of the select topiced or topics. Alternatively, fill may be set to a data vector with one entry per row of fit$L, in which case the data are mapped to the fill colour of the points. When fill = "none", the fill colour is not varied.

k

The dimensions or topics selected by number or name. When fill = "loading", one plot is created per selected dimension or topic; when fill = "loading" and k is not specified, all dimensions or topics are plotted.

fill.label

The label used for the fill colour legend.

ggplot_call

The function used to create the plot. Replace embedding_plot_2d_ggplot_call or pca_hexbin_plot_ggplot_call with your own function to customize the appearance of the plot.

plot_grid_call

When fill = "loading" and multiple topics (k) are selected, this is the function used to arrange the plots into a grid using plot_grid. It should be a function accepting a single argument, plots, a list of ggplot objects.

fill.type

The type of variable mapped to fill colour. The fill colour is not varied when fill.type = "none".

font.size

Font size used in plot.

pcs

The two principal components (PCs) to be plotted, specified by name or number.

n

The maximum number of points to plot. If n is less than the number of rows of fit$L, the rows are subsampled at random. This argument is ignored if Y is provided.

...

Additional arguments passed to pca_from_topics, tsne_from_topics or umap_from_topics. These additional arguments are only used if Y is not provided.

bins

Number of bins used to create hexagonal 2-d histogram. Passed as the “bins” argument to stat_bin_hex.

breaks

To produce the hexagonal histogram, the counts are subdivided into intervals based on breaks. Passed as the “breaks” argument to cut.

Value

A ggplot object.

Details

This is a lightweight interface primarily intended to expedite creation of plots for visualizing the loadings or topic proportions; most of the heavy lifting is done by ‘ggplot2’. The 2-d embedding itself is computed by invoking pca_from_topics, tsne_from_topics or umap_from_topics. For more control over the plot's appearance, the plot can be customized by modifying the ggplot_call and plot_grid_call arguments.

An effective 2-d visualization may also require some fine-tunning of the settings, such as the t-SNE “perplexity”, or the number of samples included in the plot. The PCA, UMAP, t-SNE settings can be controlled by the additional arguments (...). Alternatively, a 2-d embedding may be pre-computed, and passed as argument Y.

See also

Examples

set.seed(1)
data(pbmc_facs)

# Get the Poisson NMF and multinomial topic models fitted to the
# PBMC data.
fit1 <- multinom2poisson(pbmc_facs$fit)
fit2 <- pbmc_facs$fit

# Plot the first two PCs of the loadings matrix (for the
# multinomial topic model, "fit2", the loadings are the topic
# proportions).
subpop <- pbmc_facs$samples$subpop
p1 <- pca_plot(fit1,k = 1)
p2 <- pca_plot(fit2)
p3 <- pca_plot(fit2,fill = "none")
p4 <- pca_plot(fit2,pcs = 3:4,fill = "none")
p5 <- pca_plot(fit2,fill = fit2$L[,1])
p6 <- pca_plot(fit2,fill = subpop)
p7 <- pca_hexbin_plot(fit1)
p8 <- pca_hexbin_plot(fit2)

# \donttest{
# Plot the loadings using t-SNE.
p1 <- tsne_plot(fit1,k = 1)
#> Read the 2000 x 6 data matrix successfully!
#> OpenMP is working. 1 threads.
#> Using no_dims = 2, perplexity = 100.000000, and theta = 0.100000
#> Computing input similarities...
#> Building tree...
#> Done in 0.62 seconds (sparsity = 0.195510)!
#> Learning embedding...
#> Iteration 50: error is 56.566020 (50 iterations in 0.99 seconds)
#> Iteration 100: error is 49.588009 (50 iterations in 0.70 seconds)
#> Iteration 150: error is 48.780437 (50 iterations in 0.63 seconds)
#> Iteration 200: error is 48.481760 (50 iterations in 0.62 seconds)
#> Iteration 250: error is 48.321616 (50 iterations in 0.67 seconds)
#> Iteration 300: error is 0.500184 (50 iterations in 0.82 seconds)
#> Iteration 350: error is 0.355462 (50 iterations in 0.90 seconds)
#> Iteration 400: error is 0.304304 (50 iterations in 0.86 seconds)
#> Iteration 450: error is 0.280181 (50 iterations in 0.84 seconds)
#> Iteration 500: error is 0.266792 (50 iterations in 0.83 seconds)
#> Iteration 550: error is 0.258517 (50 iterations in 0.78 seconds)
#> Iteration 600: error is 0.252917 (50 iterations in 0.83 seconds)
#> Iteration 650: error is 0.248990 (50 iterations in 0.79 seconds)
#> Iteration 700: error is 0.246069 (50 iterations in 0.78 seconds)
#> Iteration 750: error is 0.243836 (50 iterations in 0.84 seconds)
#> Iteration 800: error is 0.242063 (50 iterations in 0.80 seconds)
#> Iteration 850: error is 0.240679 (50 iterations in 0.78 seconds)
#> Iteration 900: error is 0.239602 (50 iterations in 0.78 seconds)
#> Iteration 950: error is 0.238714 (50 iterations in 0.76 seconds)
#> Iteration 1000: error is 0.237994 (50 iterations in 0.74 seconds)
#> Fitting performed in 15.75 seconds.
p2 <- tsne_plot(fit2)
#> Read the 2000 x 6 data matrix successfully!
#> OpenMP is working. 1 threads.
#> Using no_dims = 2, perplexity = 100.000000, and theta = 0.100000
#> Computing input similarities...
#> Building tree...
#> Done in 0.62 seconds (sparsity = 0.185092)!
#> Learning embedding...
#> Iteration 50: error is 55.169629 (50 iterations in 1.06 seconds)
#> Iteration 100: error is 48.296536 (50 iterations in 0.67 seconds)
#> Iteration 150: error is 47.208215 (50 iterations in 0.60 seconds)
#> Iteration 200: error is 46.768378 (50 iterations in 0.56 seconds)
#> Iteration 250: error is 46.529132 (50 iterations in 0.56 seconds)
#> Iteration 300: error is 0.481751 (50 iterations in 0.72 seconds)
#> Iteration 350: error is 0.333907 (50 iterations in 0.74 seconds)
#> Iteration 400: error is 0.276316 (50 iterations in 0.72 seconds)
#> Iteration 450: error is 0.247547 (50 iterations in 0.73 seconds)
#> Iteration 500: error is 0.231735 (50 iterations in 0.69 seconds)
#> Iteration 550: error is 0.222157 (50 iterations in 0.72 seconds)
#> Iteration 600: error is 0.215846 (50 iterations in 0.67 seconds)
#> Iteration 650: error is 0.211401 (50 iterations in 0.79 seconds)
#> Iteration 700: error is 0.208026 (50 iterations in 0.65 seconds)
#> Iteration 750: error is 0.205488 (50 iterations in 0.72 seconds)
#> Iteration 800: error is 0.203476 (50 iterations in 0.68 seconds)
#> Iteration 850: error is 0.201817 (50 iterations in 0.67 seconds)
#> Iteration 900: error is 0.200466 (50 iterations in 0.67 seconds)
#> Iteration 950: error is 0.199298 (50 iterations in 0.69 seconds)
#> Iteration 1000: error is 0.198349 (50 iterations in 0.65 seconds)
#> Fitting performed in 13.94 seconds.
p3 <- tsne_plot(fit2,fill = subpop)
#> Read the 2000 x 6 data matrix successfully!
#> OpenMP is working. 1 threads.
#> Using no_dims = 2, perplexity = 100.000000, and theta = 0.100000
#> Computing input similarities...
#> Building tree...
#> Done in 0.58 seconds (sparsity = 0.184268)!
#> Learning embedding...
#> Iteration 50: error is 54.034875 (50 iterations in 1.17 seconds)
#> Iteration 100: error is 47.811956 (50 iterations in 0.73 seconds)
#> Iteration 150: error is 47.003071 (50 iterations in 0.71 seconds)
#> Iteration 200: error is 46.677254 (50 iterations in 0.67 seconds)
#> Iteration 250: error is 46.489028 (50 iterations in 0.67 seconds)
#> Iteration 300: error is 0.463680 (50 iterations in 0.72 seconds)
#> Iteration 350: error is 0.321646 (50 iterations in 0.77 seconds)
#> Iteration 400: error is 0.268448 (50 iterations in 0.82 seconds)
#> Iteration 450: error is 0.242447 (50 iterations in 0.78 seconds)
#> Iteration 500: error is 0.227653 (50 iterations in 0.77 seconds)
#> Iteration 550: error is 0.218327 (50 iterations in 0.74 seconds)
#> Iteration 600: error is 0.212016 (50 iterations in 0.78 seconds)
#> Iteration 650: error is 0.207504 (50 iterations in 0.74 seconds)
#> Iteration 700: error is 0.204139 (50 iterations in 0.75 seconds)
#> Iteration 750: error is 0.201572 (50 iterations in 0.83 seconds)
#> Iteration 800: error is 0.199485 (50 iterations in 0.78 seconds)
#> Iteration 850: error is 0.197809 (50 iterations in 0.78 seconds)
#> Iteration 900: error is 0.196472 (50 iterations in 0.79 seconds)
#> Iteration 950: error is 0.195349 (50 iterations in 0.78 seconds)
#> Iteration 1000: error is 0.194371 (50 iterations in 0.78 seconds)
#> Fitting performed in 15.54 seconds.

# Plot the loadings using UMAP.
p1 <- umap_plot(fit1,k = 1)
#> 15:47:14 UMAP embedding parameters a = 1.896 b = 0.8006
#> 15:47:14 Read 2000 rows and found 6 numeric columns
#> 15:47:14 Using FNN for neighbor search, n_neighbors = 30
#> 15:47:14 Commencing smooth kNN distance calibration using 2 threads
#> 15:47:14 Initializing from normalized Laplacian + noise
#> 15:47:14 Commencing optimization for 500 epochs, with 74134 positive edges
#> 15:47:17 Optimization finished
p2 <- umap_plot(fit2)
#> 15:47:17 UMAP embedding parameters a = 1.896 b = 0.8006
#> 15:47:17 Read 2000 rows and found 6 numeric columns
#> 15:47:17 Using FNN for neighbor search, n_neighbors = 30
#> 15:47:17 Commencing smooth kNN distance calibration using 2 threads
#> 15:47:17 64 smooth knn distance failures
#> 15:47:17 Found 3 connected components, 
#> falling back to 'spca' initialization with init_sdev = 1
#> 15:47:17 Initializing from PCA
#> 15:47:17 PCA: 2 components explained 65.66% variance
#> 15:47:17 Commencing optimization for 500 epochs, with 72866 positive edges
#> 15:47:20 Optimization finished
p3 <- umap_plot(fit2,fill = subpop)
#> 15:47:20 UMAP embedding parameters a = 1.896 b = 0.8006
#> 15:47:20 Read 2000 rows and found 6 numeric columns
#> 15:47:20 Using FNN for neighbor search, n_neighbors = 30
#> 15:47:20 Commencing smooth kNN distance calibration using 2 threads
#> 15:47:20 66 smooth knn distance failures
#> 15:47:21 Found 2 connected components, 
#> falling back to 'spca' initialization with init_sdev = 1
#> 15:47:21 Initializing from PCA
#> 15:47:21 PCA: 2 components explained 66.13% variance
#> 15:47:21 Commencing optimization for 500 epochs, with 73072 positive edges
#> 15:47:23 Optimization finished
# }