Get Poisson Non-Negative Matrix Factorization from Multinomial Topic Model

This function recovers parameter estimates of the Poisson non-negative matrix factorization given parameter estimates for a multinomial topic model.

multinom2poisson(fit, X)

Arguments

fit
An object of class “multinom_topic_model_fit”, such as an output from poisson2multinom.

X
The n x m matrix of counts or pseudocounts. It can be a sparse matrix (class "dgCMatrix") or dense matrix (class "matrix"). This only needs to be provided if the document sizes fit$s are not available.

Value

The return value is the list fit, in which matrices fit$F and fit$L specify the factors and loadings in the Poisson non-negative matrix factorization; specifically, the counts matrix is modeled by the low-rank matrix product tcrossprod(fit$L,fit$F).