PCA Plot

Visualize the structure of the Poisson NMF loadings (“activations”) or the multinomial topic proportions by projection onto principal components (PCs). plot_hexbin_plot is most useful for visualizing the PCs of a data set with thousands of samples, or more.

pca_plot(
  fit,
  k,
  out.pca,
  pcs = 1:2,
  fill = "loading",
  ggplot_call = pca_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots),
  ...
)

pca_plot_ggplot_call(
  dat,
  pcs,
  fill.type = c("loading", "numeric", "factor", "none"),
  fill.label,
  font.size = 9
)

pca_hexbin_plot(
  fit,
  out.pca,
  pcs = 1:2,
  bins = 40,
  breaks = c(0, 1, 10, 100, 1000, Inf),
  ggplot_call = pca_hexbin_plot_ggplot_call,
  ...
)

pca_hexbin_plot_ggplot_call(dat, pcs, bins, breaks, font.size = 9)

Arguments

fit
An object of class “poisson_nmf_fit” or “multinom_topic_model_fit”.

k
The topic, or topics, selected by number or name. When fill = "loading", one plot is created per selected topic; when fill = "loading" and k is not specified, all topics are plotted.

out.pca
A list containing the result of a principal components analysis, typically the result of calling prcomp or a comparable function such as rpca from the rsvd package. It should be a list object containing, at a minimum, a list matrix or data frame out.pca$x storing the rotated data. If not provided, principal components will be computed automatically by calling prcomp.

pcs
The two principal components (PCs) to be plotted, specified by name or number.

fill
The quantity to map onto the fill colour of the points in the PCA plot. Set fill = "loading" to vary the fill colour according to the loadings (or topic proportions) of the select topic, or topics. Alternatively, fill may be set to a data vector with one entry per row of fit$L, in which case these data are mapped to the fill colour of the points. When fill = "none", the fill colour is not varied.

ggplot_call
The function used to create the plot. Replace pca_plot_ggplot_call or pca_hexbin_plot_ggplot_call with your own function to customize the appearance of the plot.

plot_grid_call
When multiple topics (k) are selected, and fill = "loading", this is the function used to arrange the plots into a grid using plot_grid. It should be a function accepting a single argument, plots, a list of ggplot objects.

...
Additional arguments passed to prcomp. These additional arguments are only used if out.pca is not provided.

dat
A data frame passed as input to ggplot, containing, at a minimum, the principal components to be plotted. The data frame passed to pca_plot_ggplot_call should also have a “y” column, which is mapped to the fill colour of the points in the plot.

fill.type
The type of variable mapped to fill colour. The fill colour is not varied when fill.type = "none".

fill.label
The label used for the fill colour legend.

font.size
Font size used in plot.

bins
Number of bins used to create hexagonal 2-d histogram. Passed as the “bins” argument to stat_bin_hex.

breaks
To produce the hexagonal histogram, the counts are subdivided into intervals based on breaks. Passed as the “breaks” argument to cut.

Value

A ggplot object.

Details

This is a lightweight interface for rapidly producing PCA plots from a Poisson non-negative matrix factorization or multinomial topic model. The plotting is implemented using ggplot2. For more control over the plot's appearance, the plot can be customized by modifying the ggplot_call and plot_grid_call arguments. If not provided, the principal components are computed using prcomp.

Note that since principal components are a linear projection (“rotation”) of the data, distances between points can bee more interpretable than nonlinear embedding methods such as t-SNE (as visualized using tsne_plot) and UMAP.