Get Multinomial Topic Model from Poisson Topic Model

This function recovers parameter estimates of the multinomial topic model given parameter estimates of the Poisson topic model.

poisson2multinom(fit)

Arguments

fit
A list containing two dense, non-negative matrices, fit$F and fit$L. The former is an m x k matrix of factors, and the latter is an n x k matrix of loadings. It does not make sense for a multinomial topic model to have less than two topics, so an error will be reported when k < 2.

Value

The return value is the list fit, in which fit$F and fit$L are the parameters of the multinomial topic model; specifically, fit$L[i,] gives the topic probabilities for sample or document i, and fit$F[,k] gives the term probabilities for topic k. An additional vector fit$s of length n is returned giving the document "sizes".