Extract or Re-order Poisson NMF Factors and Loadings

This function can be used to extract estimates for a subset of the count data, or to re-order the factors and loadings (corresponding to columns and rows of the data matrix, respectively).

# S3 method for poisson_nmf_fit
select(.data, factors, loadings, ...)

Arguments

.data
Poisson NMF fit; that is, an object of class “poisson_nmf_fit”, such as an output from fit_poisson_nmf.

factors
Factor indices (names or numbers), corresponding to columns of the counts matrix. If not specified, all factors are returned.

loadings
Loadings indices (names or numbers), corresponding to rows of the counts matrix. If not specified, all loadings are returned.

...
Other arguments passed to the generic select function.

Value

A Poisson NMF fit containing the selected factors and loadings only.

Details

Evaluation metrics such as the log-likelihood and deviance are invariant to re-ordering of the factors and loadings (rows and columns of the counts matrix), but extracting a subset of the factors and/or loadings will invalidate these metrics.

See also