Structure Plot

Create a "Structure plot" from a multinomial topic model fit. The Structure plot represents the estimated topic proportions of each sample as a stacked barchart, with bars of different colors representing different topics. Consequently, samples that have similar topic proportions have similar amounts of each color.

structure_plot(
  fit,
  n = 2000,
  rows,
  grouping,
  topics,
  colors = c("#8dd3c7", "#ffffb3", "#bebada", "#fb8072", "#80b1d3", "#fdb462",
    "#b3de69", "#fccde5", "#d9d9d9", "#bc80bd", "#ccebc5", "#ffed6f"),
  gap = 1,
  ggplot_call = structure_plot_ggplot_call,
  ...
)

# S3 method for poisson_nmf_fit
plot(x, ...)

# S3 method for multinom_topic_model_fit
plot(x, ...)

structure_plot_ggplot_call(dat, colors, ticks = NULL, font.size = 9)

Arguments

fit
An object of class “poisson_nmf_fit” or “multinom_topic_model_fit”. If a Poisson NMF fit is provided as input, the corresponding multinomial topic model fit is automatically recovered using link{poisson2multinom}.

n
The maximum number of samples (rows of the loadings matrix) to include in the plot. Including large numbers (e.g., thousands) of samples is not recommended because it dramatically slows down the t-SNE computation, and typically there is little to no benefit in including large number of samples in the plot due to screen resolution limits. Ignored if rows is provided.

rows
Ordering of rows (samples) in plot, after they have been grouped. Generated automatically from 1-d t-SNE if not provided.

grouping
Optional categorical variable (factor) with one entry for each row of the loadings matrix fit$L defining a grouping of the samples (rows). The samples (rows) are arranged along the horizontal axis according to this grouping, then within each group according to rows or, if not provided, according to the 1-d t-SNE embedding.

topics
Top-to-bottom ordering of the topics in the structure plot; topics[1] is shown on the top, topics[2] is shown next, and so on. The default is topics = 1:k, where k is the number of topics (columns of fit$L).

colors
Colors used to draw topics in Structure plot: colors[1] is the colour used to draw topics[1], colors[2] is the colour used to draw topics[2], and so on. The default colour setting is the from https://colorbrewer2.org (qualitative data, "12-class Set3").

gap
The horizontal spacing between groups. Ignored if grouping is not provided.

ggplot_call
The function used to create the plot. Replace structure_plot_ggplot_call with your own function to customize the appearance of the plot.

...
Additional arguments passed to structure_plot (for the plot method) or tsne_from_topics (for function structure_plot).

x
An object of class “poisson_nmf_fit” or “multinom_topic_model_fit”. If a Poisson NMF fit is provided as input, the corresponding multinomial topic model fit is automatically recovered using link{poisson2multinom}.

dat
A data frame passed as input to ggplot, containing, at a minimum, columns "sample", "topic" and "prop": column "sample" contains the positions of the samples (rows of the loadings matrix) along the horizontal axis; column "topic" is a topic (corresponding to columns of the loadings matrix); and column "prop" is the topic proportion for the given sample.

ticks
The placement of the group labels along the horizontal axis, and their names. For data that is not grouped, use ticks = NULL.

font.size
Font size used in plot.

Value

A ggplot object.

Details

The name "Structure plot" comes from its widespread use in population genetics to visualize the results of the Structure software (Rosenberg et al, 2002).

For most uses of the Structure plot in population genetics, there is usually some grouping of the samples (e.g., assignment to pre-defined populations) that guides arrangement of the samples along the horizontal axis in the bar chart. In other applications, such as analysis of gene expression data, no pre-defined grouping exists. Therefore, a "smart" arrangement of the samples is, by default, generated automatically by performing a 1-d t-SNE embedding of the samples.

Alternatively, a categorical variable---the grouping---may be provided, in which case the samples are arranged according to that grouping, then arranged within each group using t-SNE.

References

Dey, K. K., Hsiao, C. J. and Stephens, M. (2017). Visualizing the structure of RNA-seq expression data using grade of membership models. PLoS Genetics 13, e1006599.

Rosenberg, N. A., Pritchard, J. K., Weber, J. L., Cann, H. M., Kidd, K. K., Zhivotovsky, L. A. and Feldman, M. W. (2002). Genetic structure of human populations. Science 298, 2381–2385.