Simple Interface for Fitting a Multinomial Topic Model

Fits a multinomial topic model to the count data, hiding most of the complexities of model fitting. The default optimization settings used here are intended to work well in a wide range of data sets, although some fine-tuning may be needed for more difficult cases. For full control, use fit_poisson_nmf.

fit_topic_model(
  X,
  k,
  numiter.main = 100,
  numiter.refine = 100,
  method.main = "em",
  method.refine = "scd",
  init.method = c("topicscore", "random"),
  control.init = list(),
  control.main = list(numiter = 4),
  control.refine = list(numiter = 4, extrapolate = TRUE),
  verbose = TRUE
)

Arguments

X
The n x m matrix of counts; all entries of X should be non-negative. It can be a sparse matrix (class "dgCMatrix") or dense matrix (class "matrix").

k
The number of topics. Must be 2 or greater.

numiter.main
Number of updates of the factors and loadings to perform in the main model fitting step. Should be 1 or more.

numiter.refine
Number of updates of the factors and loadings to perform in the model refinement step.

method.main
The method to use for updating the factors and loadings in the main model fitting step. Passed as argument "method" to fit_poisson_nmf.

method.refine
The method to use for updating the factors in evthe model refinement step. Passed as argument "method" to fit_poisson_nmf.

init.method
The method used to initialize the factors and loadings. See init_poisson_nmf for details.

control.init
A list of parameters controlling the behaviour of the optimization and Topic SCORE method in the call to init_poisson_nmf. This is passed as argument "control" to init_poisson_nmf.

control.main
A list of parameters controlling the behaviour of the optimization in the main model fitting step. This is passed as argument "control" to fit_poisson_nmf.

control.refine
A list of parameters controlling the behaviour of the of the optimization algorithm in the model refinement step. This is passed as argument "control" to fit_poisson_nmf.

verbose
When verbose = TRUE, information about the progress of the model fitting is printed to the console. See fit_poisson_nmf for an explanation of the output.

Value

A multinomial topic model fit; see poisson2multinom and fit_poisson_nmf for details. Note that outputted likelihoods and deviances in progress are evaluated with respect to the equivalent Poisson NMF model.

Details

With the default settings, the model fitting is accomplished in four steps: (1) initialize the Poisson NMF model fit (init_poisson_nmf); (2) perform the main model fitting step by running 100 EM updates using fit_poisson_nmf; (3) refine the fit by running 100 extrapolated SCD updates, again using fit_poisson_nmf; and (4) recover the multinomial topic model by calling poisson2multinom.

This two-stage fitting approach is based on our findings that the EM algorithm initially makes rapid progress toward a solution, but its convergence slows considerably as the iterates approach a solution. Close to a solution, we have found that other algorithms make much more rapid progress than EM; in particularly, we founr that the extrapolated SCD updates usually performed best). For larger data sets, more updates in the main model fitting and refinement steps may be needed to obtain a good fit.

For larger data sets, more than 200 updates may be needed to obtain a good fit.

References

Dey, K. K., Hsiao, C. J. and Stephens, M. (2017). Visualizing the structure of RNA-seq expression data using grade of membership models. PLoS Genetics 13, e1006599.

Blei, D. M., Ng, A. Y. and Jordan, M. I. (2003). Latent Dirichlet allocation. Journal of Machine Learning Research 3, 993-1022.

Hofmann, T. (1999). Probabilistic latent semantic indexing. In Proceedings of the 22nd International ACM SIGIR Conference, 50-57. doi:10.1145/312624.312649

See also

Examples

library(Matrix)
set.seed(1)
X <- simulate_count_data(80,100,k = 3,sparse = TRUE)$X
fit <- fit_topic_model(X,k = 3)

#> Initializing factors using Topic SCORE algorithm.
#> Initializing loadings by running 10 SCD updates.

#> Using 4 RcppParallel threads.

#> Using 4 RcppParallel threads.

#> Fitting rank-3 Poisson NMF to 80 x 100 sparse matrix.
#> Running 100 EM updates, without extrapolation (fastTopics 0.4-33).
#> iter   log-likelihood        deviance res(KKT)  |F-F'|  |L-L'| nz(F) nz(L) beta
#>    1 -8.352950339e+03 +7.79318898e+03 1.17e+00 2.6e-01 5.5e-03 0.817 0.950 0.00
#>    2 -8.346344065e+03 +7.77997644e+03 9.51e-01 1.3e-01 5.0e-02 0.817 0.950 0.00
#>    3 -8.341981803e+03 +7.77125191e+03 7.24e-01 1.0e-01 5.5e-02 0.817 0.950 0.00
#>    4 -8.339012843e+03 +7.76531399e+03 5.97e-01 6.5e-02 4.8e-02 0.817 0.950 0.00
#>    5 -8.336914655e+03 +7.76111762e+03 6.15e-01 4.5e-02 4.6e-02 0.817 0.950 0.00
#>    6 -8.335379223e+03 +7.75804675e+03 5.95e-01 3.8e-02 4.7e-02 0.817 0.950 0.00
#>    7 -8.334222177e+03 +7.75573266e+03 5.52e-01 3.1e-02 4.4e-02 0.817 0.950 0.00
#>    8 -8.333326953e+03 +7.75394221e+03 5.03e-01 2.6e-02 3.9e-02 0.817 0.950 0.00
#>    9 -8.332620114e+03 +7.75252854e+03 4.37e-01 2.2e-02 3.4e-02 0.817 0.950 0.00
#>   10 -8.332058677e+03 +7.75140566e+03 3.60e-01 1.9e-02 2.8e-02 0.813 0.950 0.00
#>   11 -8.331616745e+03 +7.75052180e+03 2.82e-01 1.7e-02 2.3e-02 0.813 0.950 0.00
#>   12 -8.331273356e+03 +7.74983502e+03 2.24e-01 1.5e-02 1.9e-02 0.813 0.950 0.00
#>   13 -8.331007681e+03 +7.74930367e+03 1.82e-01 1.3e-02 1.5e-02 0.817 0.950 0.00
#>   14 -8.330800133e+03 +7.74888857e+03 1.69e-01 1.1e-02 1.2e-02 0.817 0.950 0.00
#>   15 -8.330633728e+03 +7.74855576e+03 1.56e-01 9.0e-03 9.3e-03 0.813 0.950 0.00
#>   16 -8.330489827e+03 +7.74826796e+03 1.40e-01 8.4e-03 8.5e-03 0.813 0.942 0.00
#>   17 -8.330320819e+03 +7.74792995e+03 1.16e-01 7.7e-03 7.6e-03 0.817 0.942 0.00
#>   18 -8.329948145e+03 +7.74718460e+03 3.04e-01 1.6e-02 6.7e-03 0.817 0.938 0.00
#>   19 -8.329116018e+03 +7.74552034e+03 6.20e-01 5.4e-02 1.2e-02 0.817 0.938 0.00
#>   20 -8.328251422e+03 +7.74379115e+03 4.36e-01 7.7e-02 2.4e-02 0.817 0.938 0.00
#>   21 -8.327821832e+03 +7.74293197e+03 1.81e-01 4.5e-02 1.5e-02 0.817 0.938 0.00
#>   22 -8.327648066e+03 +7.74258444e+03 1.42e-01 1.9e-02 1.2e-02 0.817 0.938 0.00
#>   23 -8.327558958e+03 +7.74240622e+03 1.07e-01 8.5e-03 9.3e-03 0.810 0.938 0.00
#>   24 -8.327498094e+03 +7.74228450e+03 8.10e-02 5.0e-03 7.1e-03 0.813 0.938 0.00
#>   25 -8.327445975e+03 +7.74218026e+03 7.00e-02 4.5e-03 5.3e-03 0.813 0.938 0.00
#>   26 -8.327389041e+03 +7.74206639e+03 6.20e-02 4.5e-03 4.4e-03 0.813 0.933 0.00
#>   27 -8.327308158e+03 +7.74190462e+03 7.94e-02 5.3e-03 3.7e-03 0.813 0.933 0.00
#>   28 -8.327174293e+03 +7.74163689e+03 1.64e-01 1.2e-02 3.5e-03 0.813 0.933 0.00
#>   29 -8.326964786e+03 +7.74121788e+03 2.70e-01 2.4e-02 5.5e-03 0.813 0.929 0.00
#>   30 -8.326702968e+03 +7.74069424e+03 3.21e-01 3.6e-02 9.1e-03 0.813 0.929 0.00
#>   31 -8.326437382e+03 +7.74016307e+03 2.75e-01 4.0e-02 1.1e-02 0.813 0.925 0.00
#>   32 -8.326133424e+03 +7.73955516e+03 2.65e-01 3.5e-02 1.2e-02 0.813 0.925 0.00
#>   33 -8.325677809e+03 +7.73864392e+03 4.03e-01 2.7e-02 1.6e-02 0.817 0.925 0.00
#>   34 -8.325123804e+03 +7.73753591e+03 3.75e-01 3.3e-02 2.7e-02 0.817 0.925 0.00
#>   35 -8.324569047e+03 +7.73642640e+03 3.54e-01 2.9e-02 2.6e-02 0.817 0.925 0.00
#>   36 -8.324016415e+03 +7.73532114e+03 5.54e-01 5.0e-02 1.7e-02 0.820 0.921 0.00
#>   37 -8.323495185e+03 +7.73427868e+03 6.87e-01 7.3e-02 1.0e-02 0.820 0.921 0.00
#>   38 -8.323083144e+03 +7.73345460e+03 6.69e-01 8.3e-02 8.6e-03 0.820 0.921 0.00
#>   39 -8.322816566e+03 +7.73292144e+03 5.38e-01 7.6e-02 8.8e-03 0.827 0.921 0.00
#>   40 -8.322665186e+03 +7.73261868e+03 3.82e-01 5.9e-02 8.2e-03 0.827 0.917 0.00
#>   41 -8.322582141e+03 +7.73245259e+03 2.53e-01 4.1e-02 6.6e-03 0.827 0.917 0.00
#>   42 -8.322534561e+03 +7.73235743e+03 1.62e-01 2.7e-02 4.9e-03 0.827 0.917 0.00
#>   43 -8.322504734e+03 +7.73229777e+03 1.04e-01 1.8e-02 3.9e-03 0.823 0.917 0.00
#>   44 -8.322483811e+03 +7.73225593e+03 6.68e-02 1.2e-02 3.4e-03 0.827 0.917 0.00
#>   45 -8.322467089e+03 +7.73222248e+03 4.38e-02 7.7e-03 2.9e-03 0.827 0.917 0.00
#>   46 -8.322451466e+03 +7.73219124e+03 3.76e-02 5.2e-03 2.6e-03 0.827 0.917 0.00
#>   47 -8.322434075e+03 +7.73215646e+03 3.32e-02 3.6e-03 2.2e-03 0.827 0.912 0.00
#>   48 -8.322411333e+03 +7.73211097e+03 3.27e-02 2.7e-03 2.0e-03 0.827 0.908 0.00
#>   49 -8.322378018e+03 +7.73204434e+03 5.55e-02 4.6e-03 1.7e-03 0.827 0.908 0.00
#>   50 -8.322326246e+03 +7.73194080e+03 9.24e-02 7.8e-03 1.5e-03 0.823 0.904 0.00
#>   51 -8.322244582e+03 +7.73177747e+03 1.49e-01 1.3e-02 2.3e-03 0.823 0.904 0.00
#>   52 -8.322118170e+03 +7.73152465e+03 2.29e-01 2.1e-02 3.9e-03 0.827 0.904 0.00
#>   53 -8.321931464e+03 +7.73115124e+03 3.27e-01 3.1e-02 6.4e-03 0.833 0.904 0.00
#>   54 -8.321674256e+03 +7.73063682e+03 4.22e-01 4.3e-02 9.7e-03 0.833 0.904 0.00
#>   55 -8.321347227e+03 +7.72998276e+03 4.82e-01 5.4e-02 1.3e-02 0.833 0.904 0.00
#>   56 -8.320959575e+03 +7.72920746e+03 4.80e-01 6.0e-02 1.6e-02 0.833 0.904 0.00
#>   57 -8.320522083e+03 +7.72833247e+03 4.22e-01 5.8e-02 1.6e-02 0.833 0.904 0.00
#>   58 -8.320057820e+03 +7.72740395e+03 3.88e-01 4.9e-02 2.1e-02 0.837 0.904 0.00
#>   59 -8.319629901e+03 +7.72654811e+03 2.78e-01 3.8e-02 2.4e-02 0.837 0.904 0.00
#>   60 -8.319310829e+03 +7.72590996e+03 1.71e-01 2.8e-02 1.7e-02 0.837 0.904 0.00
#>   61 -8.319103336e+03 +7.72549498e+03 1.19e-01 2.0e-02 8.3e-03 0.840 0.904 0.00
#>   62 -8.318965715e+03 +7.72521974e+03 1.06e-01 1.4e-02 4.5e-03 0.843 0.900 0.00
#>   63 -8.318864580e+03 +7.72501747e+03 1.26e-01 1.3e-02 3.8e-03 0.843 0.900 0.00
#>   64 -8.318784133e+03 +7.72485657e+03 1.37e-01 1.5e-02 4.4e-03 0.843 0.900 0.00
#>   65 -8.318718811e+03 +7.72472593e+03 1.37e-01 1.6e-02 5.0e-03 0.847 0.900 0.00
#>   66 -8.318666630e+03 +7.72462157e+03 1.27e-01 1.6e-02 5.4e-03 0.847 0.896 0.00
#>   67 -8.318625871e+03 +7.72454005e+03 1.10e-01 1.5e-02 5.3e-03 0.847 0.892 0.00
#>   68 -8.318594248e+03 +7.72447680e+03 9.11e-02 1.3e-02 4.9e-03 0.847 0.892 0.00
#>   69 -8.318569223e+03 +7.72442675e+03 7.85e-02 1.0e-02 4.4e-03 0.847 0.892 0.00
#>   70 -8.318548437e+03 +7.72438518e+03 6.79e-02 8.4e-03 3.8e-03 0.843 0.887 0.00
#>   71 -8.318529887e+03 +7.72434808e+03 5.87e-02 6.7e-03 3.3e-03 0.843 0.883 0.00
#>   72 -8.318511858e+03 +7.72431202e+03 5.10e-02 5.2e-03 2.8e-03 0.843 0.883 0.00
#>   73 -8.318492746e+03 +7.72427380e+03 4.48e-02 4.1e-03 2.4e-03 0.843 0.883 0.00
#>   74 -8.318470844e+03 +7.72422999e+03 3.98e-02 3.3e-03 2.1e-03 0.840 0.883 0.00
#>   75 -8.318444164e+03 +7.72417663e+03 4.20e-02 3.4e-03 2.0e-03 0.840 0.883 0.00
#>   76 -8.318410323e+03 +7.72410895e+03 5.30e-02 4.8e-03 2.1e-03 0.840 0.883 0.00
#>   77 -8.318366569e+03 +7.72402144e+03 7.17e-02 6.6e-03 2.4e-03 0.840 0.883 0.00
#>   78 -8.318309987e+03 +7.72390828e+03 9.29e-02 8.8e-03 2.8e-03 0.840 0.883 0.00
#>   79 -8.318237927e+03 +7.72376416e+03 1.14e-01 1.1e-02 3.3e-03 0.840 0.879 0.00
#>   80 -8.318148563e+03 +7.72358543e+03 1.32e-01 1.3e-02 3.9e-03 0.840 0.875 0.00
#>   81 -8.318041536e+03 +7.72337138e+03 1.42e-01 1.5e-02 4.9e-03 0.840 0.875 0.00
#>   82 -8.317918689e+03 +7.72312568e+03 1.77e-01 1.8e-02 5.7e-03 0.840 0.875 0.00
#>   83 -8.317784614e+03 +7.72285753e+03 2.11e-01 2.2e-02 7.3e-03 0.840 0.875 0.00
#>   84 -8.317646094e+03 +7.72258049e+03 2.30e-01 2.6e-02 8.8e-03 0.840 0.875 0.00
#>   85 -8.317509908e+03 +7.72230812e+03 2.28e-01 2.8e-02 9.6e-03 0.840 0.875 0.00
#>   86 -8.317380504e+03 +7.72204932e+03 2.08e-01 2.7e-02 9.6e-03 0.840 0.875 0.00
#>   87 -8.317259703e+03 +7.72180771e+03 1.79e-01 2.5e-02 8.9e-03 0.840 0.875 0.00
#>   88 -8.317148164e+03 +7.72158463e+03 1.49e-01 2.2e-02 7.7e-03 0.840 0.875 0.00
#>   89 -8.317046540e+03 +7.72138139e+03 1.38e-01 1.8e-02 6.4e-03 0.840 0.875 0.00
#>   90 -8.316955353e+03 +7.72119901e+03 1.29e-01 1.6e-02 5.2e-03 0.840 0.875 0.00
#>   91 -8.316874306e+03 +7.72103692e+03 1.14e-01 1.5e-02 4.9e-03 0.840 0.875 0.00
#>   92 -8.316801906e+03 +7.72089212e+03 9.47e-02 1.3e-02 4.6e-03 0.840 0.875 0.00
#>   93 -8.316735517e+03 +7.72075934e+03 7.56e-02 1.0e-02 4.3e-03 0.840 0.875 0.00
#>   94 -8.316671570e+03 +7.72063145e+03 6.92e-02 8.6e-03 4.0e-03 0.840 0.875 0.00
#>   95 -8.316605849e+03 +7.72050000e+03 7.12e-02 7.9e-03 3.8e-03 0.840 0.875 0.00
#>   96 -8.316533991e+03 +7.72035629e+03 8.99e-02 8.1e-03 4.0e-03 0.840 0.875 0.00
#>   97 -8.316452511e+03 +7.72019333e+03 1.24e-01 1.2e-02 4.3e-03 0.840 0.875 0.00
#>   98 -8.316360347e+03 +7.72000900e+03 1.59e-01 1.6e-02 4.5e-03 0.840 0.875 0.00
#>   99 -8.316260155e+03 +7.71980862e+03 1.88e-01 2.0e-02 4.8e-03 0.840 0.875 0.00
#>  100 -8.316157842e+03 +7.71960399e+03 2.02e-01 2.3e-02 5.3e-03 0.840 0.871 0.00
#> Refining model fit.

#> Using 4 RcppParallel threads.

#> Fitting rank-3 Poisson NMF to 80 x 100 sparse matrix.
#> Running 100 SCD updates, with extrapolation (fastTopics 0.4-33).
#> iter   log-likelihood        deviance res(KKT)  |F-F'|  |L-L'| nz(F) nz(L) beta
#>  101 -8.315372029e+03 +7.71803236e+03 7.95e-01 1.4e-01 2.2e-02 0.853 0.858 0.55
#>  102 -8.313630585e+03 +7.71454948e+03 4.04e-01 1.8e-01 1.9e-01 0.870 0.850 0.61
#>  103 -8.312888140e+03 +7.71306459e+03 4.63e+00 1.0e-01 1.7e-01 0.867 0.833 0.67
#>  104 -8.312716137e+03 +7.71272058e+03 1.25e+00 1.2e-01 1.3e-01 0.880 0.825 0.73
#>  105 -8.312716137e+03 +7.71272058e+03 1.25e+00 4.4e-16 4.4e-16 0.880 0.825 0.55
#>  106 -8.312341271e+03 +7.71197085e+03 1.78e-01 3.1e-02 4.3e-02 0.883 0.829 0.60
#>  107 -8.312257367e+03 +7.71180304e+03 7.52e-01 4.5e-02 4.3e-02 0.880 0.829 0.66
#>  108 -8.312220738e+03 +7.71172978e+03 2.21e-01 9.0e-02 7.3e-02 0.880 0.833 0.73
#>  109 -8.312023050e+03 +7.71133441e+03 4.37e-01 1.8e-01 1.6e-01 0.880 0.829 0.80
#>  110 -8.312023050e+03 +7.71133441e+03 4.37e-01 2.2e-16 4.4e-16 0.880 0.829 0.60
#>  111 -8.311941766e+03 +7.71117184e+03 2.19e+00 2.7e-02 4.6e-02 0.873 0.817 0.66
#>  112 -8.311857937e+03 +7.71100418e+03 1.39e-01 2.3e-02 4.5e-02 0.873 0.825 0.73
#>  113 -8.311857937e+03 +7.71100418e+03 1.39e-01 2.2e-16 4.4e-16 0.873 0.825 0.55
#>  114 -8.311815726e+03 +7.71091976e+03 7.39e-02 1.2e-02 2.1e-02 0.873 0.829 0.60
#>  115 -8.311762402e+03 +7.71081311e+03 1.04e-01 2.9e-02 4.7e-02 0.877 0.829 0.66
#>  116 -8.311613366e+03 +7.71051504e+03 1.33e-01 7.2e-02 9.2e-02 0.873 0.829 0.73
#>  117 -8.311613366e+03 +7.71051504e+03 1.33e-01 2.2e-16 4.4e-16 0.873 0.829 0.55
#>  118 -8.311533277e+03 +7.71035486e+03 1.01e-01 3.7e-02 4.5e-02 0.867 0.829 0.60
#>  119 -8.311518138e+03 +7.71032458e+03 2.02e-01 5.7e-02 3.8e-02 0.863 0.825 0.66
#>  120 -8.311498459e+03 +7.71028522e+03 5.13e-02 6.3e-03 1.3e-02 0.863 0.825 0.00
#>  121 -8.311492558e+03 +7.71027342e+03 2.54e-02 5.4e-03 8.6e-03 0.863 0.825 0.73
#>  122 -8.311492311e+03 +7.71027293e+03 3.47e-02 6.7e-03 1.1e-02 0.863 0.825 0.80
#>  123 -8.311492311e+03 +7.71027293e+03 3.47e-02 2.2e-16 4.4e-16 0.863 0.825 0.60
#>  124 -8.311491249e+03 +7.71027081e+03 1.89e-02 2.7e-03 4.3e-03 0.863 0.825 0.66
#>  125 -8.311490651e+03 +7.71026961e+03 1.56e-02 4.5e-03 7.0e-03 0.863 0.825 0.73
#>  126 -8.311490651e+03 +7.71026961e+03 1.56e-02 2.2e-16 4.4e-16 0.863 0.825 0.54
#>  127 -8.311490444e+03 +7.71026919e+03 8.18e-03 1.3e-03 1.9e-03 0.863 0.825 0.60
#>  128 -8.311490364e+03 +7.71026903e+03 7.21e-03 1.4e-03 2.9e-03 0.863 0.825 0.66
#>  129 -8.311490364e+03 +7.71026903e+03 7.21e-03 0.0e+00 0.0e+00 0.863 0.825 0.49
#>  130 -8.311490314e+03 +7.71026893e+03 5.35e-03 3.5e-04 8.3e-04 0.863 0.825 0.54
#>  131 -8.311490259e+03 +7.71026882e+03 3.80e-03 8.7e-04 1.2e-03 0.863 0.825 0.60
#>  132 -8.311490227e+03 +7.71026876e+03 2.21e-03 1.4e-03 1.4e-03 0.863 0.825 0.66
#>  133 -8.311490227e+03 +7.71026876e+03 2.21e-03 2.2e-16 4.4e-16 0.863 0.825 0.49
#>  134 -8.311490214e+03 +7.71026873e+03 1.50e-03 5.6e-04 4.5e-04 0.863 0.825 0.54
#>  135 -8.311490204e+03 +7.71026871e+03 1.59e-03 6.9e-04 7.7e-04 0.863 0.825 0.60
#>  136 -8.311490202e+03 +7.71026871e+03 1.55e-03 6.6e-04 1.1e-03 0.863 0.825 0.66
#>  137 -8.311490202e+03 +7.71026871e+03 1.55e-03 2.2e-16 4.4e-16 0.863 0.825 0.49
#>  138 -8.311490199e+03 +7.71026870e+03 1.51e-03 1.7e-04 3.7e-04 0.863 0.825 0.54
#>  139 -8.311490195e+03 +7.71026870e+03 8.71e-04 2.3e-04 5.4e-04 0.863 0.825 0.60
#>  140 -8.311490194e+03 +7.71026869e+03 6.87e-04 9.9e-05 7.4e-05 0.863 0.825 0.00
#>  141 -8.311490192e+03 +7.71026869e+03 4.89e-04 1.3e-04 9.3e-05 0.863 0.825 0.66
#>  142 -8.311490191e+03 +7.71026869e+03 2.20e-04 2.3e-04 1.6e-04 0.863 0.825 0.72
#>  143 -8.311490191e+03 +7.71026869e+03 2.20e-04 0.0e+00 0.0e+00 0.863 0.825 0.54
#>  144 -8.311490191e+03 +7.71026869e+03 1.45e-04 9.1e-05 6.4e-05 0.863 0.825 0.59
#>  145 -8.311490191e+03 +7.71026869e+03 1.45e-04 2.2e-16 4.4e-16 0.863 0.825 0.45
#>  146 -8.311490191e+03 +7.71026869e+03 7.19e-05 3.9e-05 3.8e-05 0.863 0.825 0.49
#>  147 -8.311490191e+03 +7.71026869e+03 5.31e-05 3.8e-05 4.4e-05 0.863 0.825 0.54
#>  148 -8.311490191e+03 +7.71026869e+03 5.31e-05 4.4e-16 4.4e-16 0.863 0.825 0.40
#>  149 -8.311490191e+03 +7.71026869e+03 4.67e-05 3.5e-06 7.4e-06 0.863 0.825 0.45
#>  150 -8.311490191e+03 +7.71026869e+03 3.69e-05 4.5e-06 5.9e-06 0.863 0.825 0.49
#>  151 -8.311490191e+03 +7.71026869e+03 3.09e-05 9.1e-06 6.8e-06 0.863 0.825 0.54
#>  152 -8.311490191e+03 +7.71026869e+03 2.46e-05 1.2e-05 7.1e-06 0.863 0.825 0.59
#>  153 -8.311490191e+03 +7.71026869e+03 1.50e-05 1.3e-05 9.7e-06 0.863 0.825 0.65
#>  154 -8.311490191e+03 +7.71026869e+03 1.50e-05 2.2e-16 4.4e-16 0.863 0.825 0.49
#>  155 -8.311490191e+03 +7.71026869e+03 9.75e-06 2.0e-06 3.9e-06 0.863 0.825 0.54
#>  156 -8.311490191e+03 +7.71026869e+03 8.66e-06 2.0e-06 4.7e-06 0.863 0.825 0.59
#>  157 -8.311490191e+03 +7.71026869e+03 8.66e-06 2.2e-16 4.4e-16 0.863 0.825 0.44
#>  158 -8.311490191e+03 +7.71026869e+03 7.11e-06 9.1e-07 9.3e-07 0.863 0.825 0.49
#>  159 -8.311490191e+03 +7.71026869e+03 6.09e-06 1.7e-06 1.2e-06 0.863 0.825 0.54
#>  160 -8.311490191e+03 +7.71026869e+03 4.52e-06 5.7e-07 5.9e-07 0.863 0.825 0.00
#>  161 -8.311490191e+03 +7.71026869e+03 3.43e-06 7.5e-07 7.2e-07 0.863 0.825 0.59
#>  162 -8.311490191e+03 +7.71026869e+03 2.08e-06 1.3e-06 1.2e-06 0.863 0.825 0.65
#>  163 -8.311490191e+03 +7.71026869e+03 1.34e-06 1.8e-06 1.4e-06 0.863 0.825 0.71
#>  164 -8.311490191e+03 +7.71026869e+03 1.34e-06 2.2e-16 4.4e-16 0.863 0.825 0.54
#>  165 -8.311490191e+03 +7.71026869e+03 1.34e-06 4.4e-16 4.4e-16 0.863 0.825 0.59
#>  166 -8.311490191e+03 +7.71026869e+03 1.34e-06 2.2e-16 4.4e-16 0.863 0.825 0.44
#>  167 -8.311490191e+03 +7.71026869e+03 1.34e-06 0.0e+00 0.0e+00 0.863 0.825 0.49
#>  168 -8.311490191e+03 +7.71026869e+03 1.09e-06 3.7e-07 6.6e-07 0.863 0.825 0.53
#>  169 -8.311490191e+03 +7.71026869e+03 6.83e-07 2.1e-07 2.7e-07 0.863 0.825 0.59
#>  170 -8.311490191e+03 +7.71026869e+03 2.74e-07 3.1e-07 3.2e-07 0.863 0.825 0.65
#>  171 -8.311490191e+03 +7.71026869e+03 2.74e-07 4.4e-16 4.4e-16 0.863 0.825 0.49
#>  172 -8.311490191e+03 +7.71026869e+03 2.74e-07 0.0e+00 0.0e+00 0.863 0.825 0.36
#>  173 -8.311490191e+03 +7.71026869e+03 1.97e-07 9.3e-08 5.7e-08 0.863 0.825 0.40
#>  174 -8.311490191e+03 +7.71026869e+03 1.97e-07 2.2e-16 4.4e-16 0.863 0.825 0.30
#>  175 -8.311490191e+03 +7.71026869e+03 1.97e-07 0.0e+00 0.0e+00 0.863 0.825 0.23
#>  176 -8.311490191e+03 +7.71026869e+03 1.97e-07 0.0e+00 0.0e+00 0.863 0.825 0.17
#>  177 -8.311490191e+03 +7.71026869e+03 1.97e-07 0.0e+00 0.0e+00 0.863 0.825 0.13
#>  178 -8.311490191e+03 +7.71026869e+03 1.97e-07 0.0e+00 0.0e+00 0.863 0.825 0.10
#>  179 -8.311490191e+03 +7.71026869e+03 1.97e-07 0.0e+00 0.0e+00 0.863 0.825 0.07
#>  180 -8.311490191e+03 +7.71026869e+03 1.72e-07 1.6e-08 2.9e-08 0.863 0.825 0.00
#>  181 -8.311490191e+03 +7.71026869e+03 1.31e-07 1.4e-08 2.8e-08 0.863 0.825 0.08
#>  182 -8.311490191e+03 +7.71026869e+03 1.31e-07 4.4e-16 4.4e-16 0.863 0.825 0.06
#>  183 -8.311490191e+03 +7.71026869e+03 1.31e-07 2.2e-16 4.4e-16 0.863 0.825 0.06
#>  184 -8.311490191e+03 +7.71026869e+03 1.08e-07 2.4e-08 5.2e-08 0.863 0.825 0.07
#>  185 -8.311490191e+03 +7.71026869e+03 1.08e-07 2.2e-16 4.4e-16 0.863 0.825 0.05
#>  186 -8.311490191e+03 +7.71026869e+03 1.08e-07 4.4e-16 4.4e-16 0.863 0.825 0.04
#>  187 -8.311490191e+03 +7.71026869e+03 1.08e-07 2.2e-16 4.4e-16 0.863 0.825 0.03
#>  188 -8.311490191e+03 +7.71026869e+03 1.08e-07 4.4e-16 4.4e-16 0.863 0.825 0.02
#>  189 -8.311490191e+03 +7.71026869e+03 1.08e-07 2.2e-16 4.4e-16 0.863 0.825 0.02
#>  190 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.01
#>  191 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.01
#>  192 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.01
#>  193 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.01
#>  194 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.01
#>  195 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.00
#>  196 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.00
#>  197 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.00
#>  198 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.00
#>  199 -8.311490191e+03 +7.71026869e+03 1.08e-07 0.0e+00 0.0e+00 0.863 0.825 0.00
#>  200 -8.311490191e+03 +7.71026869e+03 9.61e-08 9.5e-09 2.1e-08 0.863 0.825 0.00

print(summary(fit))

#> Model overview:
#>   Number of data rows, n: 80
#>   Number of data cols, m: 100
#>   Rank/number of topics, k: 3
#> Evaluation of Poisson NMF fit (200 updates performed):
#>   Poisson NMF log-likelihood: -8.311490191185e+03
#>   Poisson NMF deviance: +7.710268688898e+03
#>   Max KKT residual: +9.610230e-08

#> Set show.size.factors = TRUE, show.mixprops = TRUE and/or show.topic.reps = TRUE in print(...) to show more informataion