Structure Plot

Create a “Structure plot” from a multinomial topic model fit. The Structure plot represents the estimated topic proportions of each sample in a stacked bar chart, with bars of different colors representing different topics. Consequently, samples that have similar topic proportions have similar amounts of each color.

structure_plot(
  fit,
  grouping,
  topics,
  rows,
  n = 2000,
  colors = c("#e41a1c", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33",
    "#a65628", "#f781bf", "#999999"),
  gap = 1,
  embed_method = structure_plot_default_embed_method,
  ggplot_call = structure_plot_ggplot_call,
  ...
)

structure_plot_default_embed_method(fit, ...)

# S3 method for poisson_nmf_fit
plot(x, ...)

# S3 method for multinom_topic_model_fit
plot(x, ...)

structure_plot_ggplot_call(dat, colors, ticks = NULL, font.size = 9)

Arguments

fit
An object of class “poisson_nmf_fit” or “multinom_topic_model_fit”. If a Poisson NMF fit is provided as input, the corresponding multinomial topic model fit is automatically recovered using poisson2multinom.

grouping
Optional categorical variable (a factor) with one entry for each row of the loadings matrix fit$L defining a grouping of the samples (rows). The samples (rows) are arranged along the horizontal axis according to this grouping, then within each group according to rows or, if not provided, according to the 1-d embedding computed by embed_method. If grouping is not a factor, an attempt is made to convert it to a factor using as.factor.

topics
Top-to-bottom ordering of the topics in the Structure plot; topics[1] is shown on the top, topics[2] is shown next, and so on. If the ordering of the topics is not specified, the topics are automatically ordered so that the topics with the greatest total “mass” are at shown at the bottom of the plot. The topics may be specified by number or by name.

rows
Ordering of the rows (samples) in the Structure plot after they have been grouped. This may include all the rows, or a subset of the rows. If not provided, it is generated automatically from a 1-d embedding, separately for each group. The rows may be specified by number or by name.

n
The maximum number of samples (rows of the loadings matrix fit$L) to include in the plot. Typically there is little to no benefit in including large number of samples in the Structure plot due to screen resolution limits. Ignored if rows is provided.

colors
Colors used to draw topics in Structure plot. The default colour setting is the from https://colorbrewer2.org (qualitative data, “9-class Set1”).

gap
The horizontal spacing between groups. Ignored if grouping is not provided.

embed_method
The function used to compute an 1-d embedding from a loadings matrix fit$L; ignored if rows is provided. The function must accept the multinomial topic model fit as its first input (“fit”) and additional arguments may be passed (...). The output should be a named numeric vector with one entry per row of fit$L, and the names of the entries should be the same as the row names of fit$L.

ggplot_call
The function used to create the plot. Replace structure_plot_ggplot_call with your own function to customize the appearance of the plot.

...
Additional arguments passed to structure_plot (for the plot method) or embed_method (for function structure_plot).

x
An object of class “poisson_nmf_fit” or “multinom_topic_model_fit”. If a Poisson NMF fit is provided as input, the corresponding multinomial topic model fit is automatically recovered using poisson2multinom.

dat
A data frame passed as input to ggplot, containing, at a minimum, columns “sample”, “topic” and “prop”: the “sample” column contains the positions of the samples (rows of the L matrix) along the horizontal axis; the “topic” column is a topic (a column of L); and the “prop” column is the topic proportion for the respective sample.

ticks
The placement of the group labels along the horizontal axis, and their names. For data that are not grouped, use ticks = NULL.

font.size
Font size used in plot.

Value

A ggplot object.

Details

The name “Structure plot” comes from its widespread use in population genetics to visualize the results of the Structure software (Rosenberg et al, 2002).

For most uses of the Structure plot in population genetics, there is usually some grouping of the samples (e.g., assignment to pre-defined populations) that guides arrangement of the samples along the horizontal axis in the bar chart. In other applications, such as analysis of gene expression data, a pre-defined grouping may not always be available. Therefore, a “smart” arrangement of the samples is, by default, generated automatically by performing a 1-d embedding of the samples.

References

Dey, K. K., Hsiao, C. J. and Stephens, M. (2017). Visualizing the structure of RNA-seq expression data using grade of membership models. PLoS Genetics 13, e1006599.

Rosenberg, N. A., Pritchard, J. K., Weber, J. L., Cann, H. M., Kidd, K. K., Zhivotovsky, L. A. and Feldman, M. W. (2002). Genetic structure of human populations. Science 298, 2381–2385.

Examples

set.seed(1)
data(pbmc_facs)

# Get the multinomial topic model fitted to the
# PBMC data.
fit <- pbmc_facs$fit

# Create a Structure plot without labels. The samples (rows) are
# automatically arranged along the x-axis using t-SNE to highlight
# the structure in the data.
p1 <- structure_plot(fit)

#> Running tsne on 2000 x 6 matrix.

# Create a Structure plot with the FACS cell-type labels. Within each
# group (cell-type), the cells (rows) are automatically arranged using
# t-SNE.
subpop <- pbmc_facs$samples$subpop
p2 <- structure_plot(fit,grouping = subpop)

#> Running tsne on 428 x 6 matrix.

#> Running tsne on 89 x 6 matrix.

#> Running tsne on 337 x 6 matrix.

#> Running tsne on 356 x 6 matrix.

#> Running tsne on 790 x 6 matrix.

# Next, we apply some customizations to improve the plot: (1) use the
# "topics" argument to specify the order in which the topic
# proportions are stacked on top of each other; (2) use the "gap"
# argrument to increase the whitespace between the groups; (3) use "n"
# to decrease the number of rows included in the Structure plot; and
# (4) use "colors" to change the colors used to draw the topic
# proportions.
topic_colors <- c("skyblue","forestgreen","darkmagenta",
                  "dodgerblue","gold","darkorange")
p3 <- structure_plot(fit,grouping = pbmc_facs$samples$subpop,gap = 20,
                     n = 1500,topics = c(5,6,1,4,2,3),colors = topic_colors)

#> Running tsne on 295 x 6 matrix.

#> Running tsne on 72 x 6 matrix.

#> Running tsne on 287 x 6 matrix.

#> Running tsne on 264 x 6 matrix.

#> Running tsne on 582 x 6 matrix.

# In this final example, we use UMAP instead of t-SNE to arrange all
# 3,744 cells in the Structure plot. Note that this can be
# accomplished in a different way by overriding the default setting of
# "embed_method".
y <- drop(umap_from_topics(fit,dims = 1))

#> 10:48:18 UMAP embedding parameters a = 1.896 b = 0.8006

#> 10:48:18 Read 3774 rows and found 6 numeric columns

#> 10:48:18 Using FNN for neighbor search, n_neighbors = 30

#> 10:48:18 Commencing smooth kNN distance calibration using 2 threads

#> 10:48:19 111 smooth knn distance failures

#> 10:48:19 Initializing from normalized Laplacian + noise

#> 10:48:20 Commencing optimization for 500 epochs, with 138012 positive edges

#> 10:48:25 Optimization finished

p4 <- structure_plot(fit,rows = order(y),grouping = subpop,gap = 40,
                     colors = topic_colors)