Volcano Plot

Create one or more "volcano" plots to visualize the results of a differential count analysis using a topic model. A volcano plot is a scatterplot in which the log-fold change (LFC), estimated using a multinomial topic model, is plotted against the p-value or z-score.

volcano_plot(
  diff_count_result,
  k,
  labels,
  y = c("zscore", "pvalue"),
  betamax = 15,
  label_above_lfc = 0,
  label_above_quantile = 0.99,
  subsample_below_quantile = 0,
  subsample_rate = 0.1,
  max.overlaps = Inf,
  ggplot_call = volcano_plot_ggplot_call,
  plot_grid_call = function(plots) do.call(plot_grid, plots)
)

# S3 method for topic_model_diff_count
plot(x, ...)

volcano_plotly(
  diff_count_result,
  k,
  file,
  labels,
  y = c("zscore", "pvalue"),
  betamax = 15,
  subsample_below_quantile = 0,
  subsample_rate = 0.1,
  width = 500,
  height = 500,
  title = paste("topic", k),
  plot_ly_call = volcano_plot_ly_call
)

volcano_plot_ggplot_call(
  dat,
  y.label,
  topic.label,
  max.overlaps = Inf,
  font.size = 9
)

volcano_plot_ly_call(dat, y.label, title, width, height)

Arguments

diff_count_result
An object of class “topic_model_diff_count”, usually an output from diff_count_analysis.

k
The topic, or topics, selected by number or name. When not specified, all topics are plotted.

labels
Character vector specifying how the points in the volcano plot are labeled. This should be a character vector with one entry per LFC estimate (row of diff_count_result$beta). When not specified, the row names of diff_count_result$beta are used, if available. Labels are added to the plot using geom_text_repel.

y
The measure of support to plot in the y-axis: use y = "zscore" to plot the z-score magnitudes; use y = "pvalue" to plot the -log1 p-values.

betamax
Truncate the LFC statistics (beta) by this amount. Any statistics greater than than betamax are set to betamax, and any statistics less than -betamax are set to -betamax.

label_above_lfc
Only z-scores or p-values (depending on choice of y) with LFC estimates above this value are labeled in the volcano plot.

label_above_quantile
Only z-scores or p-values (depending on choice of y) above this quantile are labeled in the volcano plot.

subsample_below_quantile
A number between 0 and 1. If greater than zero, LFC estimates with z-scores or p-values below this quantile will be subsampled according to subsample_rate. This can be helpful to reduce the number of points plotted for a large data set.

subsample_rate
A number between 0 and 1 giving the proportion of LFC estimates with small z-scores that are included in the plot, uniformly at random. This is only used if subsample_below_quantile is greater than zero.

max.overlaps
Argument passed to geom_text_repel.

ggplot_call
The function used to create the plot. Replace volcano_plot_ggplot_call with your own function to customize the appearance of the plot.

plot_grid_call
When multiple topics are selected, this is the function used to arrange the plots into a grid using plot_grid. It should be a function accepting a single argument, plots, a list of ggplot objects.

x
An object of class “topic_model_diff_count”, usually an output from diff_count_analysis.

...
Additional arguments passed to volcano_plot.

file
Save the interactive volcano plot to this HTML file using saveWidget.

width
Width of the plot in pixels. Passed as argument “width” to plot_ly.

height
Height of the plot in pixels. Passed as argument “height” to plot_ly.

title
The text used for the plot title.

plot_ly_call
The function used to create the plot. Replace volcano_plot_ly_call with your own function to customize the appearance of the interactive plot.

dat
A data frame passed as input to ggplot, containing, at a minimum, columns “label”, “mean”, “beta”, “se”, “z”, “pval” and “label”.

y.label
Label to use in the plot for “dat$y”.

topic.label
The name or number of the topic being plotted. Only used to determine the plot title.

font.size
Font size used in plot.

Value

A ggplot object or a plotly object.

Details

The colour of the points is varied by the average count, on the logarithmic scale; since the evidence (z-score or p-value) typically increases with more observed counts, the variables with smallest average counts should usually appear toward the bottom of the volcano plot. Only points above a specified z-score (or p-value) quantile are labeled. Note that points with an average count of zero or less are not shown.

To better accommodate situations in which some z-scores (or p-values) are much larger than all the others, the z-scores and -log10 p-values are plotted on the square-root scale. To change this, as well as other aspects, of the plot, replace volcano_plot_ggplot_call with your own function; see input argument “ggplot_call”.

The “ggrepel” package is used to arrange the labels in a visually attractive manner.

Use interactive volcano plot is created using the “plotly” package. The “hover text” shows the label (see input argument “labels”) and detailed LFC statistics as they were calculated by diff_count_analysis.